Moved cardImages and suitStyling loading  from the main layout.server.ts to cardAppearanceLoader.ts and used it only in the needed routes.
