+++ cornucopia.owasp.org/src/routes/edition/[edition]/[card]/+page.server.ts
@@ -4,6 +4,7 @@ import { DeckService } from "$lib/services/deckService";
 import type { Route } from "$domain/routes/route";
 import { MappingService } from "$lib/services/mappingService";
 import { CapecService } from "$lib/services/capecService";
+import { getCardImagesByEdition, getSuitStylingByEdition } from "$lib/services/cardAppearanceLoader";
 
 export const load = (({ params }) => {
     const edition =  params?.edition;
@@ -30,7 +31,9 @@ export const load = (({ params }) => {
       ]),
       mappingData: (new MappingService()).getCardMappingForLatestEdtions(),
       languages: DeckService.getLanguagesForEditionVersion(edition, version),
-      capecData
+      capecData,
+      cardImages: getCardImagesByEdition(),
+      suitStyling: getSuitStylingByEdition()
     };
 
     // Some QR code errors where done on the first printed decks. This will compensate for that.
