+++ cornucopia.owasp.org/src/routes/edition/[edition]/[card]/[version]/[lang]/+page.server.ts
@@ -4,6 +4,7 @@ import { error } from '@sveltejs/kit';
 import type { Route } from "$domain/routes/route";
 import { MappingService } from "$lib/services/mappingService";
 import { CapecService } from "$lib/services/capecService";
+import { getCardImagesByEdition, getSuitStylingByEdition } from "$lib/services/cardAppearanceLoader";
 
 export const load = (({ params }) => {
     const edition =  params?.edition;
@@ -41,7 +42,9 @@ export const load = (({ params }) => {
         [`${edition}`, (new MappingService()).getCardMappingForAllVersions().get(`${edition}-${version}`)]
       ]),
       languages: DeckService.getLanguagesForEditionVersion(edition, version),
-      capecData: capecData
+      capecData: capecData,
+      cardImages: getCardImagesByEdition(),
+      suitStyling: getSuitStylingByEdition()
     };
 
     // Some QR code errors where done on the first printed decks. This will compensate for that.
